Identity functional

+++ Garage 2.0/Migrations/Garage_2_0ContextModelSnapshot.cs
@@ -140,8 +140,8 @@ namespace Garage_2._0.Migrations
                     b.Property<string>("PasswordHash")
                         .HasColumnType("nvarchar(max)");
 
-                    b.Property<int>("PersonalNumber")
-                        .HasColumnType("int");
+                    b.Property<long>("PersonalNumber")
+                        .HasColumnType("bigint");
 
                     b.Property<string>("PhoneNumber")
                         .HasColumnType("nvarchar(max)");
